Upgrade to tailwind 1.x (#435)
* Update config file according to upgrade-guide

* upgrade tailwind to 1.x

* rename tailwind config file

* replace calls to config with calls to theme

* remove all use of list-reset, as this is already being done

* convert "pin" styles
