+++ addon/components/api/x-toggles/template.hbs
@@ -1,4 +1,4 @@
-<ul class="docs-list-reset docs-mb-6">
+<ul class="docs-mb-6">
   {{#each-in toggles as |key toggle|}}
     {{#if (not-eq toggle undefined)}}
       <li class="docs-inline docs-ml-4">
