+++ addon/components/docs-header/search-box/template.hbs
@@ -1,5 +1,5 @@
 <div class="docs-relative docs-text-grey-darkest" data-search-box>
-  {{svg-jar "search" width=12 height=12 class="docs-absolute docs-pin-t docs-h-full docs-ml-1"}}
+  {{svg-jar "search" width=12 height=12 class="docs-absolute docs-top-0 docs-h-full docs-ml-1"}}
 
   {{!--
     This is a dumb input - so why not use the {{input}} helper? Because apparently
