+++ addon/components/docs-header/search-results/template.hbs
@@ -7,7 +7,7 @@
     targetAttachment="bottom left"
     constraints=(array (hash to="window" attachment="together" pin=true))}}
 
-    <ul class="docs-w-76 docs-bg-white docs-shadow-md docs-list-reset" data-test-search-result-list>
+    <ul class="docs-w-76 docs-bg-white docs-shadow-md" data-test-search-result-list>
       {{#each (take 5 searchResults) as |result index|}}
         <li>
           {{docs-header/search-result
