+++ addon/components/docs-header/version-selector/template.hbs
@@ -4,7 +4,7 @@
   clickOutsideToClose=true
   onClose=(action on-close)
   targetAttachment="bottom right"}}
-  <ul class="docs-list-reset docs-mt-2 docs-bg-white docs-shadow-md docs-text-xs docs-rounded docs-z-10">
+  <ul class="docs-mt-2 docs-bg-white docs-shadow-md docs-text-xs docs-rounded docs-z-10">
     {{#each sortedVersions as |version|}}
       <li data-test-id="version">
         <a {{action "changeVersion" version}} href="#"
