+++ addon/components/docs-snippet/template.hbs
@@ -19,7 +19,7 @@
       clipboardText=snippetText
       title="copy to clipboard"
       class="
-        docs-absolute docs-pin-t docs-pin-r docs-bg-transparent hover:docs-bg-transparent
+        docs-absolute docs-top-0 docs-right-0 docs-bg-transparent hover:docs-bg-transparent
         docs-border-none docs-opacity-50 hover:docs-opacity-100 docs-p-3 docs-text-white
       "
     }}
