+++ addon/components/docs-viewer/x-current-page-index/template.hbs
@@ -1,6 +1,6 @@
 <nav class="AddonDocs-DocsViewer-CurrentPageIndex docs-hidden xl:docs-block docs-flex-no-shrink docs-mr-auto" data-test-current-page-index>
-  <div class="docs-sticky docs-pin-t docs-pt-4 docs-pb-8 docs-pr-8 docs-max-h-screen docs-overflow-y-scroll">
-    <ul class="docs-list-reset docs-border-l docs-border-grey-lighter docs-pl-6 docs-leading-normal">
+  <div class="docs-sticky docs-top-0 docs-pt-4 docs-pb-8 docs-pr-8 docs-max-h-screen docs-overflow-y-scroll">
+    <ul class="docs-border-l docs-border-grey-lighter docs-pl-6 docs-leading-normal">
       {{#if pageIndex.length}}
         <li class="docs-mt-12 docs-text-grey docs-font-bold docs-tracking-wide docs-uppercase docs-text-xxs">
           On This Page
