+++ addon/styles/addon.scss
@@ -1,7 +1,7 @@
 // Global settings
 $site-container: 1400px;
 
-@tailwind preflight;
+@tailwind base;
 @tailwind components;
 @tailwind utilities;
