+++ package.json
@@ -87,7 +87,7 @@
     "sass": "^1.22.10",
     "semver": "^6.3.0",
     "striptags": "^3.1.1",
-    "tailwindcss": "0.7.4",
+    "tailwindcss": "^1.0",
     "walk-sync": "^2.0.2",
     "yuidocjs": "^0.10.2"
   },
